# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/product": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a list of products",
                "operationId": "e68b1eace49fb57d47cb6e34def392ca",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a product",
                "operationId": "f6c6fa48346dd08aae3060e1e8765913",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Product's name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "price",
                        "in": "query",
                        "description": "Product's price",
                        "required": true,
                        "schema": {
                            "type": "double"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Product's description",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Product's category, foreign key to the category table",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "created_at",
                        "in": "query",
                        "description": "Product's creation date",
                        "required": true,
                        "schema": {
                            "type": "timestamp"
                        }
                    },
                    {
                        "name": "updated_at",
                        "in": "query",
                        "description": "Product's update date",
                        "required": true,
                        "schema": {
                            "type": "timestamp"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        }
    }
}